requirements.txt
# This file is here because many Platforms as a Service look for
#	requirements.txt in the root directory of a project.
Django
ipython
django-extensions
django-braces
django-model-utils
South
gunicorn
salt
django-uuidfield==0.5.0
django-admin-sortable
psycopg2
raven
python-simple-hipchat
